+++ src/main/java/org/nestify/backend/dto/PointMapDto.java
@@ -2,6 +2,6 @@ package org.nestify.backend.dto;
 
 public record PointMapDto(
 		String id,
-		int buildIdMapTiler
+		long buildIdMapTiler
 ) {
 }
